requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
bleach==6.2.0
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
fastapi==0.120.0
ffmpy==0.6.4
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
kaggle==1.7.4.5
keras==3.11.3
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
lxml==5.3.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.4
packaging==25.0
pandas==2.2.3
pillow==10.4.0
polars==1.34.0
polars-runtime-32==1.34.0
protobuf==5.29.5
psutil==7.1.2
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.2.0
ruff==0.14.2
safehttpx==0.1.7
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.16.2
seaborn==0.13.2
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==3.2.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.10.16
tkinter-tooltip==3.1.0
tomlkit==0.13.3
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
triton==3.5.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.221
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.38.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
